64bit実行ファイルがある程度動作するようになった
git-svn-id: https://svn.osdn.net/svnroot/ttssh2/branches/cmake@7298 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ teraterm/teraterm/ttplug.c
@@ -118,7 +118,7 @@ void WINAPI TTXInit(PTTSet ts, PComVar cv) {
   if (1) {
     char buf[1024];
     struct _finddata_t searchData;
-    long searchHandle;
+	intptr_t searchHandle;
 
     _snprintf_s(buf, sizeof(buf), _TRUNCATE, "%s\\TTX*.DLL", ts->HomeDir);
